refactor: date.js를 format.js로 변경 및 런타임 포맷 함수 통합 #17
